[Refac] Utils 리팩토링
