+++ utils/index.ts
@@ -1,4 +1,11 @@
-import { validateDate, isFuture, dateSeparate, getDateFromUrl } from "./date";
+import {
+  validateDate,
+  isFuture,
+  dateSeparate,
+  getDateFromUrl,
+  today,
+} from "./date";
+import { getFromDB, getAllFromDB, createFromDB } from "./getDB";
 import isLogin from "./login";
 
 export {
@@ -7,4 +14,8 @@ export {
   isFuture,
   isLogin,
   getDateFromUrl,
+  today,
+  getFromDB,
+  getAllFromDB,
+  createFromDB,
 };
